+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added dedicated routes for statistics and bar tabs
 - Category chips in statistics are now only shown when corresponding data is available
 - refactor statistics screen files
+- Moved logout button above roadmap info
 
 ### Fixed
 - Fix warning in web regarding missing emoji font subset